FINISHED, v1.0 of basic_logger
The project is now fully usable, no bugs known and planned features are fully available.
Since BCrypt didn't work as I intended (the hash changes for the same input, which I didn't know all along) I switch to simple AES256 encryption. However, you may want to consider NOT using this application as only client-sided anymore, but rather server-sided since the SECRET_KEY and SALTVALUE of the AES256 could be (f.e. via easy forensic analysis or code breaking) changed and this could as certain could come in handy for a hacker breaching the database as the hacker would be able to compare the hashes. Self-explanatory. I'll consider whether I'll continue working on this project or just leave it here. Neverthenless, I had great fun working on this project and it gave me a great experience in OOP and generally working with other java library to implement. Even though this project mainly served for my own interest and still wanted it to be available for everyone, I don't see the benefit anymore of putting more time into this project than it really gets acquainted attention to.